factor out enevolv writer
